+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Added: support for context as an additional of render/renderJSX
+
 # 1.0.0 - 2016-04-08
 
 ✨ Initial release